+++ CHANGELOG.md
@@ -10,6 +10,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 <!-- ### Added | Changed | Deprecated | Removed | Fixed | Security -->
 
+### Added
+
+- Options with `atoms` option, to match `yaml_elixir` library.
+
 <!--------------------- Don't add new entries after this line --------------------->
 
 ## [3.0.1] - 2022-09-05